// typings/screens-enums.ts
export enum SCREENS {
  SPLASH = 'SPLASH',
  HOME = 'HOME',
  HOMESTACK = 'HOMESTACK',
  SOCIAlARTSTACK = 'SOCIAlARTSTACK',
  WALLETSTACK = 'WALLETSTACK',
  SOCIALART = 'SOCIALART',
  APPLYEVENT = 'APPLYEVENT',
  MARKETPLACE = 'MARKETPLACE',
  EARN = 'EARN',
  ASSETS = 'ASSETS',
  TRENDING = 'TRENDING',
  ARTICLE = 'ARTICLE',
  PAYOUT = 'PAYOUT',
  MARKETPLACEDETAIL = 'MARKETPLACEDETAIL',
  CREATEMARKETPOST = 'CREATEMARKETPOST',
  CREATENEWPOSTDETAIL = 'CREATENEWPOSTDETAIL',
  CREATENEWPOSTEDIT = 'CREATENEWPOSTEDIT',
  MINTINGPOST = 'MINTINGPOST',
  PROFILE = 'PROFILE',
  CAMERA = 'CAMERA',
  TABVIEWEXAMPLE = 'TABVIEWEXAMPLE',
  CREATEPROFILE = 'CREATEPROFILE',
  DEPOSITSCREEN = 'DEPOSITSCREEN',
  TOUCHIDSETUP = 'TOUCHIDSETUP',
  TOUCHIDLOGIN = 'TOUCHIDLOGIN',
  QRCODESCANNER = 'QRCODESCANNER',
  QRAUTHORIZE = 'QRAUTHORIZE',
  WALLETSCREEN = 'WALLETSCREEN',
  WALLETDETAILS = 'WALLETDETAILS',
  WALLETCURRENCY = 'WALLETCURRENCY',
  RECOVERYSCREEN = 'RECOVERYSCREEN',
  CHOOSEAUDIENCE = 'CHOOSEAUDIENCE',
  FANSSCREEN = 'FANSSCREEN',
  CHATLISTSCREEN = 'CHATLISTSCREEN',
  CHATSCREEN = 'CHATSCREEN',
  NEWCHATSCREEN = 'NEWCHATSCREEN',
  SELLNFTDETAILS = 'SELLNFTDETAILS',
  SELL = 'SELL',
  //wallet
  SELECTNETWORK = 'SELECTNETWORK',
  TRANSACTIONHISTORY = 'TRANSACTIONHISTORY',
  WALLETSETTING = 'WALLETSETTING',
  WITHDRAWAL = 'WITHDRAWAL',
  IMPORTTOKEN = 'IMPORTTOKEN',
  ADDIMPORTACCOUNT = 'ADDIMPORTACCOUNT',
  WITHDRAWALTOKEN = 'WITHDRAWALTOKEN',
  WITHDRAWALNFT = 'WITHDRAWALNFT',
  WITHDRAWALSNFT = 'WITHDRAWALSNFT',
  RECOVERYPHRASE = 'RECOVERYPHRASE',
  STREAMTITLE = 'STREAMTITLE',
  STREAMTITLEITEM = 'STREAMTITLEITEM',
  LIVEVIDEO = 'LIVEVIDEO',
  VIDEOSTREAMITEM = 'VIDEOSTREAMITEM',
  JOINERVIDEOITEM = 'JOINERVIDEOITEM',
  LIVESTREAMEND = 'LIVESTREAMEND',
  JOINERLIVEVIDEO = 'JOINERLIVEVIDEO',
  JOINERBUY = 'JOINERBUY',
  PURCHASESTREAMITEM = 'PURCHASESTREAMITEM',
  PLACEBID = 'PLACEBID',
  AUTHENTICATEVERIFY = 'AUTHENTICATEVERIFY',
  SOCIALSTORYFEED = 'SOCIALSTORYFEED',
  TOKENTRANSACTIONHISTORY = 'TOKENTRANSACTIONHISTORY',
  POSTCOMMENTS = 'POSTCOMMENTS',
  PINAUTH = 'PINAUTH',
  LOGINPINAUTH = 'LOGINPINAUTH',
  CONFIRMPINAUTH = 'CONFIRMPINAUTH',
  MARKETPLACEDETAILSNFT = 'MARKETPLACEDETAILSNFT',
  SUBMITOFFER = 'SUBMITOFFER',
  MARKETPLACESTACK = 'MARKETPLACESTACK',
  BUYMODAL = 'BUYMODAL',
  MarketplaceFilter = 'MarketplaceFilter',
  ONBOARDINGSCREENS = 'ONBOARDINGSCREENS',
  ONBOARDINGPOST = 'ONBOARDINGPOST',
  ONBOARDINGMINT = 'ONBOARDINGMINT',
  ONBOARDINGEARNING = 'ONBOARDINGEARNING',
  ONBOARDINGEARNMORE = 'ONBOARDINGEARNMORE',
  ONBOARDINGECOSYSTEM = 'ONBOARDINGECOSYSTEM',
  SETTINGS = 'SETTINGS',
  REFERENCE = 'REFERENCE',
  PRIVACYPOLICY = 'PRIVACYPOLICY',
  HELP = 'HELP',
  FAQ = 'FAQ',
  LANGUAGE = 'LANGUAGE',
  TERMCONDITIONS = 'TERMCONDITIONS',
  EVENTS = 'EVENTS',
  TRANSACTIONAUTH = 'TRANSACTIONAUTH',
  FANSOFCARD = 'FANSOFCARD',
  PROFILEMYFANS = 'PROFILEMYFANS',
  PROFILEFANSOFCARD = 'PROFILEFANSOFCARD',
  POSTDETAILS = 'POSTDETAILS',
  NFTDETAILPAGE = 'NFTDETAILPAGE',
  NFTFIXEDPRICE = 'NFTFIXEDPRICE',
  SUCCESSFULLTRANSCTION = 'SUCCESSFULLTRANSCTION',
  BUYSNFT = 'BUYSNFT',
  NAPAVIDEOS = 'NAPAVIDEOS',
  NAPAFULLVIDEOS = 'NAPAFULLVIDEOS',
  SEARCHSCREEN = 'SEARCHSCREEN',
  RECOVERYLOGIN = "RECOVERYLOGIN",
  RECOVERYPINLOGIN = "RECOVERYPINLOGIN",
  RECOVERYCONFIRMPINLOGIN = "RECOVERYCONFIRMPINLOGIN",
  NOTIFICATIONLISTSCREEN = "NOTIFICATIONLISTSCREEN",
  NOTIFICATIONDETAILSCREEN = "NOTIFICATIONDETAILSCREEN",

}